+++ CHANGELOG.md
@@ -1,24 +1,38 @@
+# 2023-11-05
+## Version 0.07
+* Making the code more modular.
+    * Created a `CertificateManager` class that handles certificate related operations.
+    * Created a `FileManager` class that handles file operations
+    * Created a `ConnectionManager` class that handles connections to URI's.
+
+
 # 2023-08-13
 ## Version 0.06
 * Fixing python syntax and code.
+
 
 # 2023-07-09
 ## Version 0.05
 * Adding ability to allow insecure connections (`--insecure`). Useful for when a website isn't properly configured.
 
+
 # 2023-04-26
 ## Version 0.05
 * Fixing bug [no-commonname](https://github.com/TheScriptGuy/getCertificateChain/issues/9)
+
 
 # 2023-03-20
 ## Version 0.04
 * Fixing bug [MitM proxy/ssl decrypt](https://github.com/TheScriptGuy/getCertificateChain/issues/5)
 
+
 ## Version 0.03
 * Enhancement request [adding Root Found output](https://github.com/TheScriptGuy/getCertificateChain/issues/2)
 
+
 ## Version 0.02
 * Fixing bug [deprecation warning](https://github.com/TheScriptGuy/getCertificateChain/issues/1)
+
 
 # 2023-03-12
 ## Version 0.01